requirements.txt
# FastAPI Foundry - Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# RAG System (optional)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0